move some methods to utils package
